requirements.txt
fastapi
uvicorn
reportlab
openai
python-dotenv
pydantic
